luma-rag/requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
openai
tiktoken
ebooklib
beautifulsoup4
PyPDF2
together
